+++ include/solarus/hero/SwordSwingingState.h
@@ -38,7 +38,7 @@ class Hero::SwordSwingingState: public HeroState {
     bool get_can_pick_treasure(EquipmentItem& item) const override;
     bool get_can_use_shield() const override;
     bool can_sword_hit_crystal() const override;
-    bool is_cutting_with_sword(Entity& entity) override;
+    bool is_cutting_with_sword(Destructible& destructible) override;
     bool is_teletransporter_obstacle(Teletransporter& teletransporter) override;
     void notify_obstacle_reached() override;
     void notify_attacked_enemy(